+++ EGrocer.UI/EGrocer/src/app/core/rest-proxies/environments/environment.prod.ts
@@ -1,7 +1,7 @@
-import { CategoryServiceToken } from "src/app/features/categories/service/category-i.service";
-import { CategoryService } from "src/app/features/categories/service/category.service";
-import { ProductServiceToken } from "src/app/features/products/service/product-i.service";
-import { ProductService } from "src/app/features/products/service/product.service";
+import { CategoryServiceToken } from "src/app/features/category/service/category-i.service";
+import { CategoryService } from "src/app/features/category/service/category.service";
+import { ProductServiceToken } from "src/app/features/product/service/product-i.service";
+import { ProductService } from "src/app/features/product/service/product.service";
 import { FileUploadServiceToken } from "src/app/shared/service/file-upload-i.service";
 import { FileUploadService } from "src/app/shared/service/file-upload.service";
