+++ EGrocer.UI/EGrocer/src/app/core/rest-proxies/environments/environment.ts
@@ -1,7 +1,7 @@
-import { CategoryServiceToken } from "src/app/features/categories/service/category-i.service";
-import { CategoryMockService } from "src/app/features/categories/service/category.mock.service";
-import { ProductServiceToken } from "src/app/features/products/service/product-i.service";
-import { ProductMockService } from "src/app/features/products/service/product.mock.service";
+import { CategoryServiceToken } from "src/app/features/category/service/category-i.service";
+import { CategoryMockService } from "src/app/features/category/service/category.mock.service";
+import { ProductServiceToken } from "src/app/features/product/service/product-i.service";
+import { ProductMockService } from "src/app/features/product/service/product.mock.service";
 import { FileUploadServiceToken } from "src/app/shared/service/file-upload-i.service";
 import { FileUploadMockService } from "src/app/shared/service/file-upload.mock.service";
