+++ EGrocer.UI/EGrocer/src/app/features/category/details/details.component.spec.ts
