+++ EGrocer.UI/EGrocer/src/app/features/category/models/category-i.ts
