+++ EGrocer.UI/EGrocer/src/app/features/category/summary/summary.component.html
