+++ EGrocer.UI/EGrocer/src/app/features/product/details/details.component.ts
