+++ EGrocer.UI/EGrocer/src/app/features/product/models/product-i.ts
