+++ EGrocer.UI/EGrocer/src/app/features/product/summary/summary.component.spec.ts
