+++ EGrocer.UI/EGrocer/src/app/features/product/summary/summary.component.ts
